3DMASC documentation and Jupyter notebook tutorial update (#20)
* updated 3DMASC notebook with documentation and up-to-date function names. Corrected a typo in the feature_selection documentation

* added comments to the 3DMASC notebook

* updated documentation of 3DMASC jupyter notebook

+++ lidar_platform/classification/feature_selection.py
@@ -328,7 +328,7 @@ def rf_ft_selection(trads, testds, n_scales, n_features, eval_sc, threshold=0.85
         scale at which to evaluate each feature's information score at the begining of the process.
     threshold : float
         accepted value of correlation between predictors.
-    step : float
+    step : int
 
     Returns
     -------
